+++ src/routes/viewer-paths/+page.server.js
@@ -1,4 +1,4 @@
-import { getCurrentUserViewerPaths } from '$lib/server/api/auth_api.js';
+import { getCurrentUserSettings, getCurrentUserViewerPaths } from '$lib/server/api/auth_api.js';
 import { getLogger } from '$lib/server/logger.js';
 
 const logger = getLogger('viewer paths page');
@@ -7,8 +7,10 @@ export async function load({ fetch }) {
 	logger.trace('Load viewer paths page');
 
 	const viewerPaths = await getCurrentUserViewerPaths(fetch);
+	const settings = await getCurrentUserSettings(fetch);
 
 	return {
-		viewerPaths
+		viewerPaths,
+		settings
 	};
 }